添加商品修改界面

+++ Yi_Fei/ThirdLibs/Tools/PovView/SSPopup.h
@@ -25,7 +25,7 @@ typedef void (^VSActionBlock)(int);
 
 
 /* Pickup Outlet get Delegate */
--(void)GetSelectedOutlet:(int)tag;
+-(void)GetSelectedOutlet:(NSInteger)tag;
 
 
 @end
